Split trait into multiple concerns
